% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  ...
)
}
\arguments{
\item{onStart}{A function that will be called before the app is actually run.}

\item{options}{Named options that should be passed to the `shiny::runApp` call.}

\item{enableBookmarking}{Can be one of `"url"`, `"server"`, or `"disable"`.}

\item{uiPattern}{A regular expression that will be applied to each `GET` request.}

\item{...}{arguments to pass to golem_opts.
See `?golem::get_golem_options` for more details.}
}
\description{
Run the Shiny Application
}
